requirements.txt
# Required dependencies for the project
requests
pyjwt
django
lxml
signxml
dotenv
djangorestframework

# Testing dependencies (for mock server)
fastapi
uvicorn